Rename asScrollEnabled to withScrollEnabler
