+++ generatedTypes/commons/new.d.ts
@@ -1,5 +1,5 @@
 export { default as UIComponent } from './UIComponent';
 export { default as asBaseComponent, BaseComponentInjectedProps } from './asBaseComponent';
 export { default as forwardRef, ForwardRefInjectedProps } from './forwardRef';
-export { default as asScrollEnabled } from './asScrollEnabled';
+export { default as withScrollEnabler } from './withScrollEnabler';
 export { ContainerModifiers, MarginModifiers, TypographyModifiers, ColorsModifiers, BackgroundColorModifier } from './modifiers';
