+++ src/commons/index.js
@@ -22,7 +22,7 @@ module.exports = {
   get forwardRef() {
     return require('./forwardRef').default;
   },
-  get asScrollEnabled() {
-    return require('./asScrollEnabled').default;
+  get withScrollEnabler() {
+    return require('./withScrollEnabler').default;
   }
 };
