+++ src/commons/new.ts
@@ -2,7 +2,7 @@
 export {default as UIComponent} from './UIComponent';
 export {default as asBaseComponent, BaseComponentInjectedProps} from './asBaseComponent';
 export {default as forwardRef, ForwardRefInjectedProps} from './forwardRef';
-export {default as asScrollEnabled} from './asScrollEnabled';
+export {default as withScrollEnabler} from './withScrollEnabler';
 export {
   ContainerModifiers,
   MarginModifiers,
